requirements.txt
Django==1.5
South==0.7.6
django-debug-toolbar==0.9.4
pep8==1.3.3
django-jenkins==0.13.0
pyflakes==0.5.0
coverage==3.5.3
django-widget-tweaks==1.1.2
factory-boy==1.2.0
django-nose==1.1
newrelic==1.11.0.55
testfixtures==3.0.0
git+http://degit.directemployers.org/jira/jira-tools.git@57cfb93a7676ca739ce7c839038b8142a0e8a7de
